(REF) Testing - Extract class-file "TAP7"
